refactor(vscode): move the lineage canvas into the inspector panel
Relocate Canvas + graph/layout/colors/context/ContextMenu/Toolbar/nodes/overlays from panels/lineage/ into panels/inspector/ ahead of merging the two surfaces into one Rocky Inspector. Pure move at the same panel depth (src imports unchanged); the lineage entry now imports them from their new home. No behaviour change.
