+++ editors/vscode/webview-ui/panels/inspector/Canvas.tsx
