+++ editors/vscode/webview-ui/panels/inspector/ContextMenu.tsx
