+++ editors/vscode/webview-ui/panels/inspector/colors.ts
