+++ editors/vscode/webview-ui/panels/inspector/context.ts
