+++ editors/vscode/webview-ui/panels/inspector/graph.ts
