+++ editors/vscode/webview-ui/panels/inspector/layout.ts
