+++ editors/vscode/webview-ui/panels/inspector/nodes/ModelNode.tsx
