+++ editors/vscode/webview-ui/panels/inspector/overlays/freshness.ts
