+++ editors/vscode/webview-ui/panels/inspector/overlays/types.ts
